pyproject.toml
[project]
name = "backend-casadoom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-cloudauth>=0.4.3",
    "fastapi>=0.115.4",
    "pydantic[email]>=2.9.2",
    "python-dotenv>=1.0.1",
    "stytch>=11.8.0",
]
